Move RuleInformationProvider specific methods
